+++ src/main/java/entity/HoaDon.java
@@ -16,7 +16,6 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -46,13 +45,13 @@ public class HoaDon implements Serializable {
     private LocalDateTime thoiDiemTao;
 
     @Column(name = "tongTien", nullable = false, precision = 12, scale = 2)
-    private BigDecimal tongTien;
+    private double tongTien;
 
     @Column(name = "tienNhan", nullable = false, precision = 12, scale = 2)
-    private BigDecimal tienNhan;
+    private double tienNhan;
 
     @Column(name = "tienHoan", precision = 12, scale = 2)
-    private BigDecimal tienHoan;
+    private double tienHoan;
 
     @Column(name = "isThanhToanTienMat", nullable = false)
     private boolean isThanhToanTienMat;
